refactor: extract CLI agent sandbox helpers
